Refactoring - Use common functions from fsutils Version 0.4.9
