+++ src/GenesisResources.rdef
@@ -4,7 +4,7 @@ resource app_flags B_MULTIPLE_LAUNCH;
 resource app_version {
 	major  = 0,
 	middle = 4,
-	minor  = 8,
+	minor  = 9,
 
 	variety = B_APPV_DEVELOPMENT,
 	internal = 0,
